+++ public/menu_file.js
@@ -1,3 +1,4 @@
+import {MaterialHelper} from '/material_helper.js';
 import {Menu} from '/menu.js';
 import {Res} from '/res.js';
 
@@ -131,6 +132,7 @@ export class MenuFile extends Menu {
 
     updateMenuItems() {
         const res = new Res();
+        const materialHelper = new MaterialHelper();
         this.menu.items.forEach(item => {
             const data = this.itemData.items.find(
                 data => item.id == data.id
@@ -174,7 +176,14 @@ export class MenuFile extends Menu {
                     default:
                         break;
                 }
-                item.textContent = data.text + ' [' + optionText + ']';
+                materialHelper.removeChildren(item);
+                item.appendChild(materialHelper.listItemRipple());
+                item.appendChild(materialHelper.listItemText(
+                    data.text + ' [' + optionText + ']'
+                ));
+                if (data.meta) {
+                    item.appendChild(materialHelper.listItemMeta(data.meta));
+                }
             }
         });
     }